# Conflicts:
#	src/main/java/org/ton/actions/MyLocalTon.java
#	src/main/java/org/ton/executors/fift/Fift.java

+++ src/main/java/org/ton/main/App.java
@@ -1,6 +1,7 @@
 package org.ton.main;
 
 import javafx.application.Application;
+import javafx.application.Platform;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.StackPane;
@@ -25,7 +26,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.Executors;
 
-import static com.sun.javafx.PlatformUtil.isWindows;
 import static java.util.Objects.nonNull;
 
 @Slf4j
@@ -49,7 +49,17 @@ public class App extends Application {
         primaryStage.setScene(scene);
         primaryStage.setResizable(false);
         primaryStage.getScene().getWindow().addEventFilter(WindowEvent.WINDOW_CLOSE_REQUEST, this::closeWindowEvent);
-        primaryStage.setOnShown(windowEvent -> log.debug("onShown, stage loaded"));
+        primaryStage.setOnShown(windowEvent -> {
+            log.debug("onShown, stage loaded");
+
+            if (MyLocalTon.getInstance().getSettings().getActiveNodes().size() == 0) {
+                Platform.runLater(() -> mainController.showWarningMsg("Initializing TON blockchain very first time. It can take up to 2 minutes, please wait.", 60 * 5L));
+            } else if (MyLocalTon.getInstance().getSettings().getActiveNodes().size() == 1) {
+                Platform.runLater(() -> mainController.showWarningMsg("Starting TON blockchain... Should take no longer than 45 seconds.", 5 * 60L));
+            } else {
+                Platform.runLater(() -> mainController.showWarningMsg("Starting TON blockchain... Starting " + MyLocalTon.getInstance().getSettings().getActiveNodes().size() + " validators, may take up to 3 minutes.", 5 * 60L));
+            }
+        });
         primaryStage.show();
     }
 
@@ -72,12 +82,14 @@ public class App extends Application {
     public static void main(String[] args) throws Throwable {
 
         if (!Arrays.asList(args).isEmpty()) {
-            log.info("R E S E T T I N G: {}", Arrays.asList(args));
-            Thread.sleep(1000);
-            FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "MyLocalTonDB"));
-            FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "genesis"));
-            FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "templates"));
-            FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "myLocalTon.log"));
+            if (args[0].equalsIgnoreCase("restart")) {
+                log.info("R E S E T T I N G: {}", Arrays.asList(args));
+                Thread.sleep(1000);
+                FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "MyLocalTonDB"));
+                FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "genesis"));
+                FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "templates"));
+                FileUtils.deleteQuietly(new File(MyLocalTonSettings.MY_APP_DIR + File.separator + "myLocalTon.log"));
+            }
         }
 
         GraphicsEnvironment.getLocalGraphicsEnvironment().registerFont(java.awt.Font.createFont(java.awt.Font.TRUETYPE_FONT, Objects.requireNonNull(App.class.getClassLoader().getResourceAsStream("org/ton/fonts/RobotoMono-Medium.ttf"))));
@@ -89,6 +101,14 @@ public class App extends Application {
         log.info("myLocalTon config file location: {}", MyLocalTonSettings.SETTINGS_FILE);
 
         Utils.setMyLocalTonLogLevel(settings.getGenesisNode().getMyLocalTonLogLevel());
+
+        // override MyLocalTon log level
+        if (!Arrays.asList(args).isEmpty()) {
+            if (args[0].equalsIgnoreCase("debug")) {
+                Utils.setMyLocalTonLogLevel("DEBUG");
+                settings.getGenesisNode().setTonLogLevel("DEBUG");
+            }
+        }
 
         System.setProperty("objectdb.home", MyLocalTonSettings.DB_DIR);
         System.setProperty("objectdb.conf", MyLocalTonSettings.DB_SETTINGS_FILE);
@@ -117,38 +137,21 @@ public class App extends Application {
 //        dhtServer.addDhtNodesToGlobalConfig(dhtNodes2, genesisNode.getNodeGlobalConfigLocation());
 //        dhtServer.startDhtServer(settings.getNode2(), genesisNode.getNodeGlobalConfigLocation());
 
-        Thread.sleep(3000);
-
-        if (settings.getActiveNodes().size() == 1) {
-            mainController.showWarningMsg("Starting TON blockchain... Should take no longer than 45 seconds.", 5 * 60L);
-        } else {
-            mainController.showWarningMsg("Starting TON blockchain... Starting " + settings.getActiveNodes().size() + " validators, may take up to 3 minutes.", 5 * 60L);
-        }
-
         //create hardfork
         //ResultLastBlock newBlock = myLocalTon.generateNewBlock(genesisNode, forkFromBlock, "");
         //myLocalTon.addHardForkEntryIntoMyGlobalConfig(genesisNode, genesisNode.getNodeGlobalConfigLocation(), newBlock);
 
-        //before starting genesis node - let's synchronize all other nodes with it, by copying db in offline mode
-        //Utils.syncWithGenesis();
-
         new ValidatorEngine().startValidator(genesisNode, genesisNode.getNodeGlobalConfigLocation());
-
-        Thread.sleep(3000);
 
         // start other validators
         for (String nodeName : settings.getActiveNodes()) {
             if (!nodeName.contains("genesis")) {
                 long pid = new ValidatorEngine().startValidator(settings.getNodeByName(nodeName), genesisNode.getNodeGlobalConfigLocation()).pid();
                 log.info("started validator {} with pid {}", nodeName, pid);
-                if (isWindows()) {
-                    Utils.waitForBlockchainReady(settings.getNodeByName(nodeName));
-                    Utils.waitForNodeSynchronized(settings.getNodeByName(nodeName));
-                }
             }
         }
 
-        myLocalTon.runNodesMonitor();
+        myLocalTon.runNodesStatusMonitor();
 
         Utils.waitForBlockchainReady(genesisNode);
         Utils.waitForNodeSynchronized(genesisNode);
@@ -157,8 +160,9 @@ public class App extends Application {
 
         myLocalTon.runBlockchainSizeMonitor();
 
-        mainController.showSuccessMsg("TON blockchain is ready!", 2);
-        Thread.sleep(2000);
+        mainController.showSuccessMsg("TON blockchain is ready!", 3);
+
+        Thread.sleep(1000);
 
         myLocalTon.createPreInstalledWallets(genesisNode);
 
@@ -171,7 +175,7 @@ public class App extends Application {
 
         myLocalTon.runValidationMonitor();
 
-        myLocalTon.runValidatorsMonitor();
+        myLocalTon.runValidatorsProcessMonitor();
 
         mainController.addValidatorBtn.setDisable(false);
     }
